+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 === HEAD
 
+* Fix return value of 'toggle' function.
+
 === 1.0.0 (August 24, 2014)
 
 * Initial release.